requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==25.3.0
av==14.4.0
backoff==2.2.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
ctranslate2==4.6.0
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
dnspython==2.7.0
easyocr==1.7.2
email_validator==2.2.0
faiss-cpu==1.11.0
fastapi==0.115.14
fastapi-cli==0.0.7
faster-whisper==1.1.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.1
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.2
humanfriendly==10.0
idna==3.10
imageio==2.37.0
inference-sdk==0.51.1
Jinja2==3.1.6
keras==3.10.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.44.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
namex==0.1.0
networkx==3.5
ninja==1.11.1.4
numba==0.61.2
numpy==2.1.3
onnxruntime==1.22.0
openai-whisper==20250625
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pillow==11.3.0
propcache==0.3.2
protobuf==5.29.5
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pyreadline3==3.5.4
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.8
scikit-image==0.25.2
scipy==1.16.0
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
supervision==0.25.1
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-hub==0.16.1
termcolor==3.1.0
tf_keras==2.19.0
tifffile==2025.6.11
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.1